added front end validation for user registration.  moved model to models folder
