+++ models/User.js
@@ -1,0 +1,24 @@
+var Sequelize = require('sequelize');
+if (process.env.NODE_ENV === 'production') {
+  // HEROKU DB
+  console.log(process.env.JAWSDB_URL);
+
+  var connection = new Sequelize(process.env.JAWSDB_URL);
+} else {
+  // LOCAL DB
+  var connection = new Sequelize('mmdb', 'root');
+}
+
+var User = connection.define('User', {
+  email: {
+    type: Sequelize.STRING,
+    unique: true
+  },
+  password: Sequelize.STRING,
+  firstname: Sequelize.STRING,
+  lastname: Sequelize.STRING,
+  phone:Sequelize.STRING
+
+});
+
+module.exports = User;
