+++ views/signUp.handlebars
@@ -2,20 +2,25 @@
 <form method="POST" action="/register">
       <div class="form-group">
         <label for="exampleInputEmail1">Email address</label>
-        <input type="email" class="form-control" id="exampleInputEmail1" placeholder="Email" name="email">
+        <input required type="email" class="form-control" id="exampleInputEmail1" placeholder="Email" name="email">
       </div>
       <div class="form-group">
         <label for="exampleInputPassword1">Password</label>
-        <input type="password" class="form-control" id="exampleInputPassword1" placeholder="Password" name="password">
+        <input required type="password" class="form-control"  placeholder="Password" name="password">
       </div>
       <div class="form-group">
         <label for="exampleInputPassword1">First Name</label>
-        <input type="text" class="form-control" id="exampleInputPassword1" name="firstname">
+        <input required type="text" class="form-control"  name="firstname">
       </div>
 
       <div class="form-group">
         <label for="exampleInputPassword1">Last Name</label>
-        <input type="text" class="form-control" id="exampleInputPassword1" name="lastname">
+        <input required type="text" class="form-control"  name="lastname">
+      </div>
+      <div class="form-group">
+         <div class="form-group">
+        <label for="exampleInputPassword1">Phone Number</label>
+        <input required placeholder="000-000-0000" type="text" class="form-control" pattern="^\d{3}-\d{3}-\d{4}$" name="phone">
       </div>
 
       <button type="submit" class="btn btn-default">Submit</button>
