+++ views/test.handlebars
@@ -1,0 +1,5 @@
+
+<form>
+ <input pattern="^\d{3}-\d{3}-\d{4}$">
+ <button type="submit">
+  </form>
